--- a/Changelog.md
+++ b/Changelog.md
@@ -3,6 +3,7 @@
 ## v0.6.1 (unreleased)
 
 - The lower boundary for packets included in ACKs is now derived, and the value sent in STOP_WAITING frames is ignored.
+- Implement packet pacing.
 
 ## v0.6.0 (2017-12-12)