--- a/ackhandler/sent_packet_handler.go
+++ b/ackhandler/sent_packet_handler.go
@@ -364,24 +364,22 @@ func (h *sentPacketHandler) GetStopWaitingFrame(force bool) *wire.StopWaitingFra
 	return h.stopWaitingManager.GetStopWaitingFrame(force)
 }
 
-func (h *sentPacketHandler) TimeUntilSend(now time.Time) time.Duration {
-	// check if we're limited by the number of packets we're tracking
-	if protocol.PacketNumber(len(h.retransmissionQueue)+h.packetHistory.Len()) >= protocol.MaxTrackedSentPackets {
-		return utils.InfDuration
+func (h *sentPacketHandler) SendingAllowed() bool {
+	cwnd := h.congestion.GetCongestionWindow()
+	congestionLimited := h.bytesInFlight > cwnd
+	maxTrackedLimited := protocol.PacketNumber(len(h.retransmissionQueue)+h.packetHistory.Len()) >= protocol.MaxTrackedSentPackets
+	if congestionLimited {
+		utils.Debugf("Congestion limited: bytes in flight %d, window %d", h.bytesInFlight, cwnd)
 	}
-	delay := h.congestion.TimeUntilSend(now, h.bytesInFlight)
-	congestionLimited := delay == utils.InfDuration
 	// Workaround for #555:
 	// Always allow sending of retransmissions. This should probably be limited
 	// to RTOs, but we currently don't have a nice way of distinguishing them.
 	haveRetransmissions := len(h.retransmissionQueue) > 0
-	if haveRetransmissions && congestionLimited {
-		return 0 // don't pace retransmissions
-	}
-	if congestionLimited {
-		utils.Debugf("Congestion limited: bytes in flight %d, window %d", h.bytesInFlight, h.congestion.GetCongestionWindow())
-	}
-	return delay
+	return !maxTrackedLimited && (!congestionLimited || haveRetransmissions)
+}
+
+func (h *sentPacketHandler) TimeUntilSend(now time.Time) time.Duration {
+	return h.congestion.TimeUntilSend(now, h.bytesInFlight)
 }
 
 func (h *sentPacketHandler) retransmitOldestTwoPackets() {