--- a/congestion/cubic_sender.go
+++ b/congestion/cubic_sender.go
@@ -56,6 +56,9 @@ type cubicSender struct {
 	// ACK counter for the Reno implementation.
 	congestionWindowCount protocol.ByteCount
 
+	// time that the last packet was sent
+	lastPacketSentTime time.Time
+
 	initialCongestionWindow    protocol.PacketNumber
 	initialMaxCongestionWindow protocol.PacketNumber
 }
@@ -76,15 +79,24 @@ func NewCubicSender(clock Clock, rttStats *RTTStats, reno bool, initialCongestio
 	}
 }
 
+// TimeUntilSend returns when the next packet should be sent.
+// When congestion limited, it returns an infinite duration (utils.InfDuration).
 func (c *cubicSender) TimeUntilSend(now time.Time, bytesInFlight protocol.ByteCount) time.Duration {
 	if c.InRecovery() {
 		// PRR is used when in recovery.
 		return c.prr.TimeUntilSend(c.GetCongestionWindow(), bytesInFlight, c.GetSlowStartThreshold())
 	}
-	if c.GetCongestionWindow() > bytesInFlight {
+	if c.GetCongestionWindow() <= bytesInFlight {
+		return utils.InfDuration
+	}
+	delay := c.rttStats.SmoothedRTT() / time.Duration(2*c.GetCongestionWindow()/protocol.DefaultTCPMSS)
+	if !c.InSlowStart() { // adjust delay, such that it's 1.25*cwd/rtt
+		delay = delay * 8 / 5
+	}
+	if c.lastPacketSentTime.IsZero() {
 		return 0
 	}
-	return utils.InfDuration
+	return utils.MaxDuration(0, c.lastPacketSentTime.Add(delay).Sub(now))
 }
 
 func (c *cubicSender) OnPacketSent(sentTime time.Time, bytesInFlight protocol.ByteCount, packetNumber protocol.PacketNumber, bytes protocol.ByteCount, isRetransmittable bool) bool {
@@ -97,6 +109,7 @@ func (c *cubicSender) OnPacketSent(sentTime time.Time, bytesInFlight protocol.By
 		c.prr.OnPacketSent(bytes)
 	}
 	c.largestSentPacketNumber = packetNumber
+	c.lastPacketSentTime = sentTime
 	c.hybridSlowStart.OnPacketSent(packetNumber)
 	return true
 }