EDIT: I found some flaws in this, and I'm working on a fix. No need to review this PR at the current moment.

Fixes #355.

This PR adds send pacing according to the IETF Loss Recovery draft.
The most important change is in `session.sendPacket`. This function now only sends at most one packet at a time. Furthermore, it now has two return values, one bool indicating if a packet was sent, and the error. **Only if** a packet was sent, we now set the timer to the next time a packet should be sent according to the packet packer. This way, if we have a lot of data to send, we'll send it at the desired pacing rate. If there's no data to send, we avoid going through the run loop every single time.

## Benchmark results

The benchmarks were performed by having Chrome upload a 50 MB file to 1. a TCP/HTTPS server, a quic-go server and by having a quic-go client upload to a quic-go server.

without packet pacing:
```
+-----------------+----------------------------+--------------------------+--------------------------+
|                 | Chrome => TCP/HTTPS server | Chrome => quic-go server | quic-go client => server |
+-----------------+----------------------------+--------------------------+--------------------------+
| direct transfer | 156.35 ± 37.83             | 63.25 ± 8.66             | 109.16 ± 2.89            |
| 10ms RTT        | 69.50 ± 4.62               | 34.71 ± 3.82             | 12.14 ± 0.91             |
| 25ms RTT        | 30.17 ± 1.86               | 21.19 ± 7.02             | 4.06 ± 0.44              |
| 50ms RTT        | 15.44 ± 0.47               | 13.79 ± 1.18             | 2.17 ± 0.11              |
| 100ms RTT       | 7.84 ± 0.34                | 7.30 ± 0.11              | 1.62 ± 0.35              |
+-----------------+----------------------------+--------------------------+--------------------------+
Based on 4 samples of 50 MB. All values in MB/s.
```

with packet pacing:

```
+-----------------+----------------------------+--------------------------+--------------------------+
|                 | Chrome => TCP/HTTPS server | Chrome => quic-go server | quic-go client => server |
+-----------------+----------------------------+--------------------------+--------------------------+
| direct transfer | 120.58 ± 41.98             | 68.50 ± 7.22             | 23.03 ± 1.48             |
| 10ms RTT        | 70.38 ± 4.39               | 32.91 ± 5.87             | 22.39 ± 0.52             |
| 25ms RTT        | 27.98 ± 1.21               | 19.62 ± 3.54             | 19.77 ± 1.54             |
| 50ms RTT        | 14.18 ± 0.81               | 14.16 ± 0.28             | 13.71 ± 0.33             |
| 100ms RTT       | 8.95 ± 0.42                | 6.93 ± 0.86              | 6.88 ± 0.19              |
+-----------------+----------------------------+--------------------------+--------------------------+
Based on 4 samples of 50 MB. All values in MB/s.
```

Packet pacing dramatically improves the performance for all conditions except for 0 RTT. I'm planning to do some detailed benchmarking of what is going on for the 0 RTT case once this PR (and the PRs leading up to #1081) are merged. I think we should merge this first, since it brings an improvement to basically all realistic use cases.