@marten-seemann 2017-12-27
@lucas-clemente: I fixed the PR. All tests are passing now. From my side, this is ready for merging now.